+++ pyisy/constants.py
@@ -173,6 +173,7 @@ METHOD_COMMAND = "cmd"
 METHOD_GET = "get"
 METHOD_SET = "set"
 
+URL_CHANGE = "change"
 URL_CLOCK = "time"
 URL_CONFIG = "config"
 URL_DEFINITIONS = "definitions"